[ty] Move `ScopedUseId` to `UseDefMap`
